requirements.txt
Django==2.1.2
django-ckeditor==5.7.1
django-cors-headers==3.1.0
django-js-asset==1.2.2
djangorestframework==3.8.2
Pillow==6.1.0
pytz==2019.2
uWSGI==2.0.18
